% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usn.R
\name{deduplicate_usn}
\alias{deduplicate_usn}
\title{Remove duplicates (by date and wall)}
\usage{
deduplicate_usn(x = read_usn())
}
\arguments{
\item{x}{sf MUTLIPOINT table}
}
\value{
the input with duplicates dropped
}
\description{
We just keep the first occruence for each duplicate
}
