% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{diff_patch_month}
\alias{diff_patch_month}
\title{Compute differences of patch_month data}
\usage{
diff_patch_month(x, what = "median")
}
\arguments{
\item{x}{tibble of class "patch_month"}

\item{what}{chr, the name of the variable to plot}
}
\value{
tibble with cold - warm difference
}
\description{
Compute differences of patch_month data
}
