% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi-package.R
\docType{package}
\name{egad}
\alias{gsi}
\alias{gsi-package}
\title{egad package}
\description{
R-language tools for workign with Guld Stream data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bigelowlab/gsi}
}

}
