% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_points.R
\name{estimate_max_distance}
\alias{estimate_max_distance}
\title{Given a set of s2 vertices, estimate the max search distance}
\usage{
estimate_max_distance(vertices, p = 0.1)
}
\arguments{
\item{vertices}{either something that can be cast to sf's POINT or s2's wk_xy}

\item{p}{num, the qauntile as a probability - \code{\link[stats]{quantile}}}
}
\description{
Given a set of s2 vertices, estimate the max search distance
}
