% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{extract_points_usn}
\alias{extract_points_usn}
\title{Given one or more formatted strings, retrieve a matrix of lon,lat}
\usage{
extract_points_usn(
  x = c("26.5N79.8W 26.7N79.8W 26.9N79.8W 27.0N79.8W 27.2N79.8W 27.3N79.9W",
    "27.4N79.9W 27.5N80.0W 27.6N80.0W 27.8N80.1W 28.0N80.1W 28.1N80.2W")
)
}
\arguments{
\item{char, }{one or more character strings of point data as scraped from the URL}
}
\value{
matrix of point in lon, lat order
}
\description{
Given one or more formatted strings, retrieve a matrix of lon,lat
}
