% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_date}
\alias{get_date}
\title{Extract data out of the text line with date info}
\usage{
get_date(x)
}
\arguments{
\item{x}{cahr, the line of text with the date}
}
\value{
Date class object
}
\description{
Extract data out of the text line with date info
}
