% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsgi.R
\name{import_gsgi}
\alias{import_gsgi}
\title{Import (from matlab) the GSGI index data}
\usage{
import_gsgi(
  file =
    "/mnt/s1/projects/ecocast/coredata/gstream/gsi/NOAA_OISST_GS_dSST_index_monthly_1993_2019_release.mat"
)
}
\arguments{
\item{file}{char, the name of the file to import}
}
\value{
a tibble of data
}
\description{
Import (from matlab) the GSGI index data
}
