% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usn.R
\name{order_usn}
\alias{order_usn}
\title{Convert a MULTIPOINT to a LINESTRING (or MULTIPOINT) but an attempt to order
from lower left (SW) to upper right (NE)}
\usage{
order_usn(
  x = usn_example(ordered = FALSE),
  type = c("LINESTRING", "MULTIPOINT")[1],
  direction = c("any", "eastward", "westward")[2],
  resolve_multiline = c("keep", "longest")[1]
)
}
\arguments{
\item{x}{a MULTIPOINT object}

\item{type}{char, one of "LINESTRING" (default) or "MULTIPOINT"}

\item{direction}{chr one of "any", "eastward" (default), "westward"}

\item{resolve_multiline}{char, when \code{type} is LINESTRING a conflict may arise when
a wall is composed of multiple lines, which yields MULTILINESTRING.
Set this to keep just the longest LINESTRING or to keep them all as MULTILINESTRING.}
}
\value{
a LINESTRING object
}
\description{
Convert a MULTIPOINT to a LINESTRING (or MULTIPOINT) but an attempt to order
from lower left (SW) to upper right (NE)
}
