% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi.R
\name{plot.gsi}
\alias{plot.gsi}
\title{Plot gsi data time series}
\usage{
\method{plot}{gsi}(x, smooth = TRUE, by = c("none", "month", "year")[1])
}
\arguments{
\item{x}{the gsi data}

\item{smooth}{logical, if TRUE add a smoothing line (for timeseries only)}

\item{by}{chr, one of 'none' (default timeseries), 'month' (ghost plot by month over the years)
or 'year' a box plot by year}
}
\value{
a ggplot2 object
}
\description{
Plot gsi data time series
}
