% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{plot.patch_month}
\alias{plot.patch_month}
\title{Plot patch_month data}
\usage{
\method{plot}{patch_month}(x = read_patch_month(), what = "median")
}
\arguments{
\item{x}{tibble of class "patch_month"}

\item{what}{chr, the name of the variable to plot}
}
\value{
ggplot2 object
}
\description{
Plot patch_month data
}
