% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapid.R
\name{plot.rapid_moc_transports}
\alias{plot.rapid_moc_transports}
\title{Plot the moc_transports data set}
\usage{
\method{plot}{rapid_moc_transports}(x = read_moc_transports())
}
\arguments{
\item{x}{moc_transports data frame (class of "rapid_moc_transports")}
}
\value{
ggplot2 object
}
\description{
Plot the moc_transports data set
}
