% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsgi.R
\name{plot_gsgi}
\alias{plot_gsgi}
\title{Plot gsgi data time series}
\usage{
plot_gsgi(x, smooth = TRUE)
}
\arguments{
\item{x}{the gsgi data}

\item{smooth}{logical, if TRUE add a smoothing line}
}
\value{
a ggplot2 object
}
\description{
Plot gsgi data time series
}
