% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{plot_patch_location}
\alias{plot_patch_location}
\title{Plot the patches}
\usage{
plot_patch_location(x = read_patch_bbs(), what = c("cold_blob", "warm_spot"))
}
\arguments{
\item{x}{sf table of patches}

\item{what}{chr, the patches to include in the plot}
}
\value{
invisible NULL
}
\description{
Plot the patches
}
