% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_configuration}
\alias{read_configuration}
\title{Read configuration}
\usage{
read_configuration(filename = "~/.gstream")
}
\arguments{
\item{filename}{char, the config file (with path)}
}
\value{
configuration list
}
\description{
Read configuration
}
