% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsgi.R
\name{read_gsgi}
\alias{read_gsgi}
\title{Read the GS SST gradient index}
\usage{
read_gsgi(file = system.file("gsgi/gsgi.csv.gz", package = "gstream"))
}
\arguments{
\item{filename}{char, the name of the file to read}
}
\value{
tibble of date, deseasoned SST north and south and the difference (gradient)
}
\description{
Read the GS SST gradient index
}
