% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi.R
\name{read_gsi}
\alias{read_gsi}
\title{Import GSI data from the \href{https://noaa-edab.github.io/ecodata/index.html}{ecoodata} package.}
\usage{
read_gsi()
}
\value{
tibble pf class "gsi"
}
\description{
Import GSI data from the \href{https://noaa-edab.github.io/ecodata/index.html}{ecoodata} package.
}
