% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapid.R
\name{read_moc_transports}
\alias{read_moc_transports}
\title{Read the moc_transport data file}
\usage{
read_moc_transports(
  filename = system.file("rapid-amoc/moc_transports.nc", package = "gstream")
)
}
\arguments{
\item{filename}{the name of the file to read}
}
\value{
a tibble of class "rapid_moc_transports"
}
\description{
Read the moc_transport data file
}
