% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{read_patch_month}
\alias{read_patch_month}
\title{Read the patch data}
\usage{
read_patch_month(
  filename = system.file("patches/patches_data_month.csv.gz", package = "gstream")
)
}
\arguments{
\item{filename}{chr, the name of the file to read}
}
\value{
tibble of class "patch_month"
}
\description{
Read the patch data
}
