% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usn.R
\name{read_usn}
\alias{read_usn}
\title{Read one or more US Navy wall data files}
\usage{
read_usn(year = "all", what = "ordered", deduplicate = (what != "ordered"))
}
\arguments{
\item{year}{num or char, one or more years to read, or "all" for reading them
all.}

\item{what}{char, if "ordered" then read the ordered data, or "orig" for unordered.}

\item{deduplicate}{logical, if TRUE remove duplicates by date-wall}
}
\value{
SF multipoint table
}
\description{
Read one or more US Navy wall data files
}
