% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_wall_data}
\alias{read_wall_data}
\title{Get the GS wall data}
\usage{
read_wall_data(filename, verbose = FALSE)
}
\arguments{
\item{filename}{char, the name of file}

\item{verbose}{logical, if TRUE output messages}
}
\value{
tibble or sf MULTIPOINT object
}
\description{
Get the GS wall data
}
