% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_wall_data_navy}
\alias{read_wall_data_navy}
\title{Read the GS wall data in navy format}
\usage{
read_wall_data_navy(filename, verbose = FALSE)
}
\arguments{
\item{filename}{char, the name of file, a vector of files, or the nameof a
single tar file (with one or more *.sub files within)}

\item{verbose}{logical, if TRUE output messages}
}
\value{
tibble or sf MULTIPOINT object
}
\description{
Read the GS wall data in navy format
}
