% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_points.R
\name{sf_as_s2lnglat}
\alias{sf_as_s2lnglat}
\title{Convert an sf object to a s2_lnglat object}
\usage{
sf_as_s2lnglat(x = read_usn(what = "orig"))
}
\arguments{
\item{x}{an SF object to convert}
}
\value{
list of wk_xy object (one per row of x)
}
\description{
Convert an sf object to a s2_lnglat object
}
