% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{st_path_lengths}
\alias{st_path_lengths}
\title{Compute distances between consecutive points along a LINESTRING path.}
\usage{
st_path_lengths(
  x = dplyr::filter(read_usn(year = "2020"), date == as.Date("2020-12-19"), wall ==
    "north"),
  ...
)
}
\arguments{
\item{x}{sf LINESTRING object}
}
\value{
a list of path lengths, one per row
}
\description{
Compute distances between consecutive points along a LINESTRING path.
}
