% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usn.R
\name{usn_example}
\alias{usn_example}
\title{Retrieve a prepared example file of the north wall}
\usage{
usn_example(
  filename = system.file("examples/2020-12-19-north.gpkg", package = "gstream"),
  reorder = TRUE
)
}
\arguments{
\item{filename}{char, the name of the file}

\item{reorder}{logical, if TRUE try to reorder the points}
}
\value{
sf table
}
\description{
Retrieve a prepared example file of the north wall
}
