% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usn.R
\name{usn_to_network}
\alias{usn_to_network}
\title{Convert USN LINESTRING features to sfnetworks}
\usage{
usn_to_network(x = usn_example())
}
\arguments{
\item{x}{sf LINESTRING table}
}
\value{
list of networks, one per input row
}
\description{
Convert USN LINESTRING features to sfnetworks
}
