% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usn.R
\name{walls_to_polygons}
\alias{walls_to_polygons}
\title{Convert daily north-south wall pairs to a closed polygon (hopefully one!)}
\usage{
walls_to_polygons(x, bb = NULL)
}
\arguments{
\item{x}{sf table of wall data.  Only matched north-south pairs for
each given data are converted. All others are dropped.}

\item{bb}{bounding box (or object from which we can get a bounding box) to crop to
or NULL to skip}

\item{sf}{table of polygons by date}
}
\description{
Convert daily north-south wall pairs to a closed polygon (hopefully one!)
}
