% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_year}
\alias{fetch_year}
\title{Fetch a single file}
\usage{
fetch_year(
  year = "NavoGS_2010.tar",
  base_uri = "https://ftp.opc.ncep.noaa.gov/grids/experimental/GStream"
)
}
\arguments{
\item{base_uri}{char, the base uri}

\item{filename}{the name of the file}
}
\value{
sf MULTIPOINT data frame
}
\description{
Fetch a single file
}
