% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi-package.R
\docType{package}
\name{gstream}
\alias{gstream}
\alias{gstream-package}
\title{gstream package}
\description{
R-language tools for working with Gulf Stream data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bigelowlab/gstream}
}

}
