% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usn.R
\name{order_usn}
\alias{order_usn}
\title{Convert a MULTIPOINT to a LINESTRING (or MULTIPOINT) but an attempt to order
from lower left (SW) to upper right (NE)}
\usage{
order_usn(x = read_usn(year = 2020), type = c("LINESTRING", "MULTIPOINT")[1])
}
\arguments{
\item{x}{a MULTIPOINT object}

\item{type}{char, one of "LINESTRING" (default) or "MULTIPOINT"}
}
\value{
a LINESTRING object
}
\description{
Convert a MULTIPOINT to a LINESTRING (or MULTIPOINT) but an attempt to order
from lower left (SW) to upper right (NE)
}
