% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usn.R
\name{walls_to_polygons}
\alias{walls_to_polygons}
\title{Convert daily north-south wall pairs to a closed polygon (hopefully one!)}
\usage{
walls_to_polygons(
  x,
  bb = NULL,
  verbose = FALSE,
  include = c("area", "centroid"),
  use_s2 = FALSE
)
}
\arguments{
\item{x}{sf table of wall data.  Only matched north-south pairs for
each given data are converted. All others are dropped.}

\item{bb}{bounding box (or object from which we can get a bounding box) to crop to
or NULL to skip}

\item{include}{char vector of desired elements to add.  Only "area" and "centroid" are known}

\item{use_s2}{logical, if TRUE then turn on use of the s2 engine}
}
\value{
sf table of polygons by date
}
\description{
Convert daily north-south wall pairs to a closed polygon (hopefully one!)
}
