% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usn.R
\name{walls_to_polygons}
\alias{walls_to_polygons}
\title{Convert daily north-south wall pairs to a closed polygon (hopefully one!)}
\usage{
walls_to_polygons(x)
}
\arguments{
\item{x}{sf table of wall data.  Only matched north-south pairs for
each given daya are converted. All others are dropped.}

\item{sf}{table of polygons by date}
}
\description{
Convert daily north-south wall pairs to a closed polygon (hopefully one!)
}
